Updated exampaper.py
Made the script more user friendly

+++ exampaper.py
@@ -3,8 +3,9 @@
 # Ubuntu 18.04
 # Visual Studio Code
 # Python 3.6.9
+# Script Written by Szymon Masternak 2020.
 
-import requests, csv, json, io, sys, os, shutil, argparse, urllib.request
+import requests, csv, json, io, sys, os, shutil, argparse, urllib.request, datetime
 
 def progress(count, total, status=''):
     bar_len = 20
@@ -22,40 +23,59 @@ def stringtofloat(stringwithfloat):
     except ValueError:
         return 0.0
 
-parser = argparse.ArgumentParser(description='Download Batch DCU Exam papers')
+parser = argparse.ArgumentParser(description="Download Batch DCU Exam papers.", epilog="Example Usage:\"./exampaper.py -m FSH104,EE206 -y 2016 -r yes -d ./exampapersem2/\" Downloads modules fsh104 and ee206 up to year 2016 including repeat exams in the directory exampapersem2")
 parser.add_argument("-m", required=True, help="Input Your Modules Seperated by a comma")
 parser.add_argument("-r", help="Boolean statement if you want to include repeat exams. Default: Yes", default="Yes")
-parser.add_argument("-d", help="Directory You Want to Download Files to", default="./exampapers/")
-parser.add_argument("-y", help="Specify the year you want to go back to", default="0")
-parser.add_argument("-v", help="Verbose, if you want to include debug messages. Default: False", default="no")
+parser.add_argument("-d", help="Directory You Want to Download Files to. Defaults to current directory", default="./")
+parser.add_argument("-y", help="Specify the year limit you want to go back to", default="0")
+parser.add_argument("-v", help="Verbose Mode. Default: Off", action="store_true")
 
-try:
-    options = parser.parse_args()
-    module_arr = options.m
-    repeat_boolean_string = options.r
-    dir = options.d
-    year = float(options.y)
-    debug = options.v
-except:
+if len(sys.argv[1:])==0:
     parser.print_help()
-    sys.exit(0)
+    parser.exit()
+
+options = parser.parse_args()
+module_arr = options.m
+repeat_boolean_string = options.r
+dirinput = options.d
+year = float(options.y)
+debug = options.v
+
+module_arr = module_arr.lower()
 
 if str.lower(repeat_boolean_string) == "y" or str.lower(repeat_boolean_string) == "yes" or str.lower(repeat_boolean_string) == "t" or str.lower(repeat_boolean_string) == "true" or str.lower(repeat_boolean_string) == "on":
     repeat_bool = True
 else:
     repeat_bool = False
 
-if str.lower(debug) == "y" or str.lower(debug) == "yes" or str.lower(debug) == "t" or str.lower(debug) == "true" or str.lower(debug) == "on":
-    debug = True
-else:
-    debug = False
+todaysdate = datetime.datetime.now()
+
+if not dirinput.endswith("/"):
+    dirinput = dirinput + "/"
+
+dir = dirinput + "exampapers" + todaysdate.strftime("%d%m%y") + "/"
+
+if debug == True:
+    print("The Directory is  to place the file is" + dir)
 
 if os.path.isdir(dir):
-    print("Folder Found Deleting previous folder")
-    try:
-        shutil.rmtree(dir)
-    except:
-        print('Error while deleting directory')
+    print("Previous downloaded directory found. Press yes to delete it or no to exit the program.")
+    userchoice = input()
+    
+    if str.lower(userchoice) == "y" or str.lower(userchoice) == "yes" or str.lower(userchoice) == "t" or str.lower(userchoice) == "true" or str.lower(userchoice) == "on":
+        userchoice_bool = True
+    else:
+        userchoice_bool = False
+    
+    if userchoice_bool == True:        
+        try:
+            shutil.rmtree(dir)
+        except:
+            print('Error while deleting directory')
+            sys.exit(0)
+    else:
+        print("Exiting the Program. Goodbye")
+        sys.exit(0)
 
 if debug == True:
     print("m argument is " + module_arr)
@@ -88,13 +108,13 @@ for z in module_list:
     for i in range(0, length_of_list):
         if repeat_bool == True:
             if stringtofloat(data[i][1]) >= year:
-                if z in str(data[i][0]):   
+                if z in str(data[i][0]).lower():   
                     if debug == True:
                         print(data[i][0] + "  " + data[i][1]+ "  " + data[i][2])
                     counter = counter + 1
         else:
             if stringtofloat(data[i][1]) >= year:
-                if str(data[i][0]) == z:
+                if str(data[i][0]).lower() == z:
                     if debug == True:
                         print(data[i][0] + "  " + data[i][1]+ "  " + data[i][2])
                     counter = counter + 1
@@ -112,7 +132,7 @@ for z in module_list:
     for i in range(0, length_of_list):
         if repeat_bool == True:
             if stringtofloat(data[i][1]) >= year:
-                if z in str(data[i][0]):
+                if z in str(data[i][0]).lower():
                     if os.path.isdir(dir+data[i][0]) == False:
                         try:
                             os.makedirs(dir+data[i][0])
@@ -131,7 +151,7 @@ for z in module_list:
                         print("Failed to Download " + data[i][0] + "_" + data[i][1] + ".pdf")
 
         else:
-            if str(data[i][0]) == z:
+            if str(data[i][0]).lower == z:
                 if stringtofloat(data[i][1]) >= year:
                     if os.path.isdir(dir+data[i][0]) == False:
                         try:
@@ -149,9 +169,4 @@ for z in module_list:
                         urllib.request.urlretrieve(URL, dir+data[i][0]+"/"+data[i][0]+"_"+data[i][1]+".pdf")
                     except URLERROR:
                         print("Failed to Download " + data[i][0] + "_" + data[i][1] + ".pdf")
-
-                # else:
-                #     print("directory already exists lol")
-
-
